requirements.txt
# Core dependencies
pyautogen>=0.2.0
google-generativeai>=0.3.0
python-telegram-bot>=20.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0
mcp>=1.0.0

# Web UI dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0

# Additional utilities
pydantic>=2.0.0
typing-extensions>=4.5.0
loguru>=0.7.0
httpx>=0.24.0

# Super ModuMentor enhancements
Pillow>=10.0.0
schedule>=1.2.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
